requirements.txt
# Baselines & APIs
openai

# Experimentation & Analysis
hydra-core
pandas
seaborn
matplotlib
jupyter

# Data & Model Versioning
dvc[s3]

# SonarQube Communication
requests

# Utilities
python-dotenv

tqdm